Add files via upload
This Python script simulates the rolling of one or more dice, providing a simple and effective way to generate random outcomes. It can be used for games, probability experiments, or just for fun. The script allows you to specify the number of dice and the number of sides on each die, making it highly customizable for various scenarios.